Move EventBusSubscriber into outer class
