+++ loader/src/main/java/net/neoforged/fml/javafmlmod/Mod.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) Forge Development LLC and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
+package net.neoforged.fml.javafmlmod;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * This defines a Mod to FML.
+ * Any class found with this annotation applied will be loaded as a Mod. The instance that is loaded will
+ * represent the mod to other Mods in the system. It will be sent various subclasses of {@code ModLifecycleEvent}
+ * at pre-defined times during the loading of the game.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Mod {
+    /**
+     * The unique mod identifier for this mod.
+     * <b>Required to be lowercased in the english locale for compatibility. Will be truncated to 64 characters long.</b>
+     *
+     * This will be used to identify your mod for third parties (other mods), it will be used to identify your mod for registries such as block and item registries.
+     * By default, you will have a resource domain that matches the modid. All these uses require that constraints are imposed on the format of the modid.
+     */
+    String value();
+}
